+++ lib/datafoodconsortium/connector/person.rb
@@ -24,6 +24,7 @@ require "datafoodconsortium/connector/agent"
 
 
 
+
 require "virtual_assembly/semantizer"
 
 class DataFoodConsortium::Connector::Person < DataFoodConsortium::Connector::Agent
@@ -36,21 +37,22 @@ class DataFoodConsortium::Connector::Person < DataFoodConsortium::Connector::Age
 	# @return [String]
 	attr_accessor :lastName
 
-	# @return [IEnterprise]
+	# @return [IOrganization]
 	attr_accessor :affiliatedOrganizations
 
 	# @param semanticId [String]
 	# @param firstName [String]
 	# @param lastName [String]
-	# @param affiliatedOrganizations [IEnterprise]
+	# @param affiliatedOrganizations [IOrganization]
 	# @param localizations [IAddress]
 	# @param phoneNumbers [IPhoneNumber]
 	# @param emails [String]
 	# @param websites [String]
 	# @param socialMedias [ISocialMedia]
 	# @param logo [String]
-	def initialize(semanticId, firstName: nil, lastName: nil, affiliatedOrganizations: [], localizations: [], phoneNumbers: [], emails: [], websites: [], socialMedias: [], logo: nil)
-		super(semanticId, localizations: localizations, phoneNumbers: phoneNumbers, emails: emails, websites: websites, socialMedias: socialMedias, logo: logo)
+	# @param customerCategoriesMembership [ICustomerCategory]
+	def initialize(semanticId, firstName: nil, lastName: nil, affiliatedOrganizations: [], localizations: [], phoneNumbers: [], emails: [], websites: [], socialMedias: [], logo: nil, customerCategoriesMembership: [])
+		super(semanticId, localizations: localizations, phoneNumbers: phoneNumbers, emails: emails, websites: websites, socialMedias: socialMedias, logo: logo, customerCategoriesMembership: customerCategoriesMembership)
 		@firstName = firstName
 		@lastName = lastName
 		@affiliatedOrganizations = affiliatedOrganizations