+++ lib/datafoodconsortium/connector/physical_place.rb
@@ -22,6 +22,7 @@
 
 
 
+
 require "virtual_assembly/semantizer"
 
 class DataFoodConsortium::Connector::PhysicalPlace
@@ -45,10 +46,10 @@ class DataFoodConsortium::Connector::PhysicalPlace
 	attr_accessor :openingHours
 
 	# @return [IAddress]
-	attr_accessor :addresses
+	attr_accessor :address
 
 	# @return [IPerson]
-	attr_accessor :mainContact
+	attr_accessor :mainContacts
 
 	# @return [ITheoreticalStock]
 	attr_accessor :theoreticalStocks
@@ -56,37 +57,43 @@ class DataFoodConsortium::Connector::PhysicalPlace
 	# @return [IRealStock]
 	attr_accessor :realStocks
 
+	# @return [IGeoJsonFeature]
+	attr_accessor :features
+
 	# @param semanticId [String]
 	# @param name [String]
 	# @param description [String]
 	# @param hostedSaleSessions [ISaleSession]
 	# @param phoneNumbers [IPhoneNumber]
 	# @param openingHours [IOpeningHoursSpecification]
-	# @param addresses [IAddress]
-	# @param mainContact [IPerson]
+	# @param address [IAddress]
+	# @param mainContacts [IPerson]
 	# @param theoreticalStocks [ITheoreticalStock]
 	# @param realStocks [IRealStock]
-	def initialize(semanticId, name: nil, description: nil, hostedSaleSessions: [], phoneNumbers: [], openingHours: [], addresses: [], mainContact: nil, theoreticalStocks: [], realStocks: [])
+	# @param features [IGeoJsonFeature]
+	def initialize(semanticId, name: nil, description: nil, hostedSaleSessions: [], phoneNumbers: [], openingHours: [], address: nil, mainContacts: [], theoreticalStocks: [], realStocks: [], features: [])
 		super(semanticId)
 		@name = name
 		@description = description
 		@hostedSaleSessions = hostedSaleSessions
 		@phoneNumbers = phoneNumbers
 		@openingHours = openingHours
-		@addresses = addresses
-		@mainContact = mainContact
+		@address = address
+		@mainContacts = mainContacts
 		@theoreticalStocks = theoreticalStocks
 		@realStocks = realStocks
+		@features = features
 		self.semanticType = "dfc-b:PhysicalPlace"
 		registerSemanticProperty("dfc-b:name", &method("name")).valueSetter = method("name=")
 		registerSemanticProperty("dfc-b:description", &method("description")).valueSetter = method("description=")
 		registerSemanticProperty("dfc-b:hosts", &method("hostedSaleSessions")).valueSetter = method("hostedSaleSessions=")
 		registerSemanticProperty("dfc-b:hasPhoneNumber", &method("phoneNumbers")).valueSetter = method("phoneNumbers=")
 		registerSemanticProperty("dfc-b:hasOpeningHours", &method("openingHours")).valueSetter = method("openingHours=")
-		registerSemanticProperty("dfc-b:hasAddress", &method("addresses")).valueSetter = method("addresses=")
-		registerSemanticProperty("dfc-b:hasMainContact", &method("mainContact")).valueSetter = method("mainContact=")
+		registerSemanticProperty("dfc-b:hasAddress", &method("address")).valueSetter = method("address=")
+		registerSemanticProperty("dfc-b:hasMainContact", &method("mainContacts")).valueSetter = method("mainContacts=")
 		registerSemanticProperty("dfc-b:localizes", &method("theoreticalStocks")).valueSetter = method("theoreticalStocks=")
 		registerSemanticProperty("dfc-b:stores", &method("realStocks")).valueSetter = method("realStocks=")
+		registerSemanticProperty("dfc-b:hasGeoJsonFeature", &method("features")).valueSetter = method("features=")
 	end