+++ lib/datafoodconsortium/connector/realized_production_flow.rb
@@ -47,7 +47,7 @@ class DataFoodConsortium::Connector::RealizedProductionFlow < DataFoodConsortium
 		@transformation = transformation
 		@product = product
 		self.semanticType = "dfc-b:AsRealizedProductionFlow"
-		registerSemanticProperty("dfc-b:outcomeOf", &method("transformation")).valueSetter = method("transformation=")
+		registerSemanticProperty("dfc-b:outputOf", &method("transformation")).valueSetter = method("transformation=")
 		registerSemanticProperty("dfc-b:produces", &method("product")).valueSetter = method("product=")
 	end