+++ lib/datafoodconsortium/connector/variant.rb
@@ -33,23 +33,19 @@ require "datafoodconsortium/connector/defined_product"
 
 require "virtual_assembly/semantizer"
 
-class DataFoodConsortium::Connector::SuppliedProduct < DataFoodConsortium::Connector::DefinedProduct
+class DataFoodConsortium::Connector::Variant < DataFoodConsortium::Connector::DefinedProduct
 
-    SEMANTIC_TYPE = "dfc-b:SuppliedProduct".freeze
+    SEMANTIC_TYPE = "dfc-b:Variant".freeze
 
-	# @return [Real]
-	attr_accessor :totalTheoreticalStock
+	# @return [IDefinedProduct]
+	attr_accessor :isVariantOf
 
-	# @return [ILocalizedProduct]
-	attr_accessor :localizedProducts
-
-	
-
-	
+	# @return [IVariantCharacteristic]
+	attr_accessor :variantCharacteristics
 
 	# @param semanticId [String]
-	# @param totalTheoreticalStock [Real]
-	# @param localizedProducts [ILocalizedProduct]
+	# @param isVariantOf [IDefinedProduct]
+	# @param variantCharacteristics [IVariantCharacteristic]
 	# @param name [String]
 	# @param description [String]
 	# @param productType [ISKOSConcept]
@@ -69,13 +65,13 @@ class DataFoodConsortium::Connector::SuppliedProduct < DataFoodConsortium::Conne
 	# @param images [String]
 	# @param variants [IDefinedProduct]
 	# @param referenceProductOptions [IProductOption]
-	def initialize(semanticId, totalTheoreticalStock: nil, localizedProducts: [], name: nil, description: nil, productType: nil, quantity: nil, alcoholPercentage: nil, lifetime: nil, claims: [], usageOrStorageConditions: nil, allergenCharacteristics: [], nutrientCharacteristics: [], physicalCharacteristics: [], geographicalOrigin: nil, catalogItems: [], certifications: [], natureOrigin: [], partOrigin: [], images: [], variants: [], referenceProductOptions: [])
+	def initialize(semanticId, isVariantOf: [], variantCharacteristics: [], name: nil, description: nil, productType: nil, quantity: nil, alcoholPercentage: nil, lifetime: nil, claims: [], usageOrStorageConditions: nil, allergenCharacteristics: [], nutrientCharacteristics: [], physicalCharacteristics: [], geographicalOrigin: nil, catalogItems: [], certifications: [], natureOrigin: [], partOrigin: [], images: [], variants: [], referenceProductOptions: [])
 		super(semanticId, name: name, description: description, productType: productType, quantity: quantity, alcoholPercentage: alcoholPercentage, lifetime: lifetime, claims: claims, usageOrStorageConditions: usageOrStorageConditions, allergenCharacteristics: allergenCharacteristics, nutrientCharacteristics: nutrientCharacteristics, physicalCharacteristics: physicalCharacteristics, geographicalOrigin: geographicalOrigin, catalogItems: catalogItems, certifications: certifications, natureOrigin: natureOrigin, partOrigin: partOrigin, images: images, variants: variants, referenceProductOptions: referenceProductOptions)
-		@totalTheoreticalStock = totalTheoreticalStock
-		@localizedProducts = localizedProducts
-		self.semanticType = "dfc-b:SuppliedProduct"
-		registerSemanticProperty("dfc-b:totalTheoreticalStock", &method("totalTheoreticalStock")).valueSetter = method("totalTheoreticalStock=")
-		registerSemanticProperty("dfc-b:referenceOf", &method("localizedProducts")).valueSetter = method("localizedProducts=")
+		@isVariantOf = isVariantOf
+		@variantCharacteristics = variantCharacteristics
+		self.semanticType = "dfc-b:Variant"
+		registerSemanticProperty("dfc-b:isVariantOf", &method("isVariantOf")).valueSetter = method("isVariantOf=")
+		registerSemanticProperty("dfc-b:hasVariantCaracteristic", &method("variantCharacteristics")).valueSetter = method("variantCharacteristics=")
 	end